requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
aiosqlite>=0.19.0
Pillow>=10.0.0
pytz>=2023.3
flask>=2.3.0
psutil>=5.9.0
